Moving the Store_model class to avoid circular import
